+++ talkinghead/tha3/app/manual_poser.py
@@ -59,7 +59,7 @@ import os
 import pathlib
 import sys
 import time
-from typing import Dict, List, Tuple
+from typing import List
 
 import PIL.Image
 
@@ -71,8 +71,8 @@ import wx
 
 from tha3.poser.modes.load_poser import load_poser
 from tha3.poser.poser import Poser, PoseParameterCategory, PoseParameterGroup
-from tha3.util import rgba_to_numpy_image, grid_change_to_numpy_image, \
-    rgb_to_numpy_image, resize_PIL_image, extract_PIL_image_from_filelike, extract_pytorch_image_from_PIL_image
+from tha3.util import resize_PIL_image, extract_PIL_image_from_filelike, extract_pytorch_image_from_PIL_image
+from tha3.app.util import load_emotion_presets, posedict_to_pose, pose_to_posedict, torch_image_to_numpy, FpsStatistics
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -94,97 +94,6 @@ logger = logging.getLogger(__name__)
 # output_ext_to_index = {ext: idx for idx, ext in enumerate(output_index_to_ext)}
 # input_exts_and_descs_str = "|".join(format_fileformat_list(PIL_supported_input_formats))  # filter-spec accepted by `wx.FileDialog`
 # output_exts_and_descs_str = "|".join(format_fileformat_list(PIL_supported_output_formats))
-
-# The keys for a pose in the emotion JSON files.
-#
-# TODO: "eye_unimpressed" is arity-2, but has only one entry in the JSON. The current implementation smashes both into one,
-#       letting the second one (right slider) win. Maybe the two values should be saved separately, but we have to avoid
-#       breaking the live mode served by `app.py`.
-posedict_keys = ["eyebrow_troubled_left_index", "eyebrow_troubled_right_index",
-                 "eyebrow_angry_left_index", "eyebrow_angry_right_index",
-                 "eyebrow_lowered_left_index", "eyebrow_lowered_right_index",
-                 "eyebrow_raised_left_index", "eyebrow_raised_right_index",
-                 "eyebrow_happy_left_index", "eyebrow_happy_right_index",
-                 "eyebrow_serious_left_index", "eyebrow_serious_right_index",
-                 "eye_wink_left_index", "eye_wink_right_index",
-                 "eye_happy_wink_left_index", "eye_happy_wink_right_index",
-                 "eye_surprised_left_index", "eye_surprised_right_index",
-                 "eye_relaxed_left_index", "eye_relaxed_right_index",
-                 "eye_unimpressed", "eye_unimpressed",
-                 "eye_raised_lower_eyelid_left_index", "eye_raised_lower_eyelid_right_index",
-                 "iris_small_left_index", "iris_small_right_index",
-                 "mouth_aaa_index",
-                 "mouth_iii_index",
-                 "mouth_uuu_index",
-                 "mouth_eee_index",
-                 "mouth_ooo_index",
-                 "mouth_delta",
-                 "mouth_lowered_corner_left_index", "mouth_lowered_corner_right_index",
-                 "mouth_raised_corner_left_index", "mouth_raised_corner_right_index",
-                 "mouth_smirk",
-                 "iris_rotation_x_index", "iris_rotation_y_index",
-                 "head_x_index", "head_y_index",
-                 "neck_z_index",
-                 "body_y_index", "body_z_index",
-                 "breathing_index"]
-assert len(posedict_keys) == 45
-
-
-def load_emotion_presets() -> Tuple[Dict[str, Dict[str, float]], List[str]]:
-    """Load emotion presets from disk.
-
-    These are JSON files in "talkinghead/emotions".
-
-    Returns the tuple `(emotions, emotion_names)`, where::
-
-        emotions = {emotion0_name: posedict0, ...}
-        emotion_names = [emotion0_name, emotion1_name, ...]
-
-    The dict contains the actual pose data. The list is a sorted list of emotion names
-    that can be used to map a linear index (e.g. the choice index in a GUI dropdown)
-    to the corresponding key of `emotions`.
-
-    The directory "talkinghead/emotions" must also contain a "_defaults.json" file,
-    containing factory defaults (as a fallback) for the 28 standard emotions
-    (as recognized by distilbert), as well as a hidden "zero" preset that represents
-    a neutral pose. (This is separate from the "neutral" emotion, which is allowed
-    to be "non-zero".)
-    """
-    emotion_names = []
-    for root, dirs, files in os.walk("emotions", topdown=True):
-        for filename in files:
-            if filename == "_defaults.json":  # skip the repository containing the default fallbacks
-                continue
-            if filename.lower().endswith(".json"):
-                emotion_names.append(filename[:-5])  # drop the ".json"
-    emotion_names.sort()  # the 28 actual emotions
-
-    # TODO: Note that currently, we build the list of emotion names from JSON filenames,
-    #       and then check whether each JSON implements the emotion matching its filename.
-    #       On second thought, I'm not sure whether that makes much sense. Maybe rethink the design.
-    #         - We *do* want custom JSON files to show up in the list, if those are placed in "tha3/emotions". So the list of emotions shouldn't be hardcoded.
-    #         - *Having* a fallback repository with factory defaults (and a hidden "zero" preset) is useful.
-    #           But we are currently missing a way to reset an emotion to its factory default.
-    def load_emotion_with_fallback(emotion_name: str) -> Dict[str, float]:
-        try:
-            with open(os.path.join("emotions", f"{emotion_name}.json"), "r") as json_file:
-                emotions_from_json = json.load(json_file)  # A single json file may contain presets for multiple emotions.
-            posedict = emotions_from_json[emotion_name]
-        except (FileNotFoundError, KeyError):  # If no separate json exists for the specified emotion, load the default (all 28 emotions have a default).
-            with open(os.path.join("emotions", "_defaults.json"), "r") as json_file:
-                emotions_from_json = json.load(json_file)
-            posedict = emotions_from_json[emotion_name]
-        # If still not found, it's an error, so fail-fast: let the app exit with an informative exception message.
-        return posedict
-
-    # Dict keeps its keys in insertion order, so define some special states before inserting the actual emotions.
-    emotions = {"[custom]": {},  # custom = the user has changed at least one value manually after last loading a preset
-                "[reset]": load_emotion_with_fallback("zero")}  # reset = a preset with all sliders in their default positions. Found in "_defaults.json".
-    for emotion_name in emotion_names:
-        emotions[emotion_name] = load_emotion_with_fallback(emotion_name)
-
-    emotion_names = list(emotions.keys())
-    return emotions, emotion_names
 
 
 class SimpleParamGroupsControlPanel(wx.Panel):
@@ -404,43 +313,6 @@ class MorphCategoryControlPanel(wx.Panel):
         self.update_ui()
 
 
-def convert_output_image_from_torch_to_numpy(output_image):
-    if output_image.shape[2] == 2:
-        h, w, c = output_image.shape
-        numpy_image = torch.transpose(output_image.reshape(h * w, c), 0, 1).reshape(c, h, w)
-    elif output_image.shape[0] == 4:
-        numpy_image = rgba_to_numpy_image(output_image)
-    elif output_image.shape[0] == 3:
-        numpy_image = rgb_to_numpy_image(output_image)
-    elif output_image.shape[0] == 1:
-        c, h, w = output_image.shape
-        alpha_image = torch.cat([output_image.repeat(3, 1, 1) * 2.0 - 1.0, torch.ones(1, h, w)], dim=0)
-        numpy_image = rgba_to_numpy_image(alpha_image)
-    elif output_image.shape[0] == 2:
-        numpy_image = grid_change_to_numpy_image(output_image, num_channels=4)
-    else:
-        raise RuntimeError(f"Unsupported # image channels: {output_image.shape[0]}")
-    numpy_image = numpy.uint8(numpy.rint(numpy_image * 255.0))
-    return numpy_image
-
-
-class FpsStatistics:
-    def __init__(self):
-        self.count = 100
-        self.fps = []
-
-    def add_fps(self, fps: float) -> None:
-        self.fps.append(fps)
-        while len(self.fps) > self.count:
-            del self.fps[0]
-
-    def get_average_fps(self) -> float:
-        if len(self.fps) == 0:
-            return 0.0
-        else:
-            return sum(self.fps) / len(self.fps)
-
-
 class MyFileDropTarget(wx.FileDropTarget):
     def OnDropFiles(self, x, y, filenames):
         if len(filenames) > 1:
@@ -571,7 +443,7 @@ class MainFrame(wx.Frame):
         self.left_panel_sizer.Add(self.source_image_panel, 0, wx.FIXED_MINSIZE)
 
         # Emotion picker.
-        self.emotions, self.emotion_names = load_emotion_presets()
+        self.emotions, self.emotion_names = load_emotion_presets("emotions")
 
         # # Horizontal emotion picker layout; looks bad, text label vertical alignment is wrong.
         # self.emotion_panel = wx.Panel(self.left_panel, style=wx.SIMPLE_BORDER, size=(-1, -1))
@@ -767,7 +639,7 @@ class MainFrame(wx.Frame):
             if len(emotions_from_json) > 1:
                 logger.warning(f"File {json_file_name} contains multiple emotions, loading the first one '{first_emotion_name}'.")
             posedict = emotions_from_json[first_emotion_name]
-            pose = self.posedict_to_pose(posedict)
+            pose = posedict_to_pose(posedict)
 
             # Apply loaded emotion
             self.set_current_pose(pose)
@@ -863,7 +735,7 @@ class MainFrame(wx.Frame):
             emotion_name = self.emotion_choice.GetString(current_emotion_index)
             logger.info(f"Loading emotion preset {emotion_name}")
             posedict = self.emotions[emotion_name]
-            pose = self.posedict_to_pose(posedict)
+            pose = posedict_to_pose(posedict)
             self.set_current_pose(pose)
             current_pose = pose
         else:
@@ -913,7 +785,7 @@ class MainFrame(wx.Frame):
                 with torch.no_grad():
                     output_image = self.poser.pose(self.torch_source_image, pose, output_index)[0].detach().cpu()
 
-                numpy_image = convert_output_image_from_torch_to_numpy(output_image)
+                numpy_image = torch_image_to_numpy(output_image)
                 self.last_output_numpy_image = numpy_image
                 wx_image = wx.ImageFromBuffer(
                     numpy_image.shape[0],
@@ -962,25 +834,6 @@ class MainFrame(wx.Frame):
                     self.update_in_progress = False
                 wx.CallAfter(update_images_cont2)
         wx.CallAfter(update_images_cont)
-
-    def current_pose_to_posedict(self) -> Dict[str, float]:
-        """Convert the character's current pose into a posedict for saving into an emotion JSON."""
-        current_pose_values = self.get_current_pose()
-        current_pose_dict = dict(zip(posedict_keys, current_pose_values))
-        return current_pose_dict
-
-    def posedict_to_pose(self, posedict: Dict[str, float]) -> List[float]:
-        """Convert a posedict (from an emotion JSON) into a list of morph values (in the order the models expect them)."""
-        # sanity check
-        unrecognized_keys = set(posedict.keys()) - set(posedict_keys)
-        if unrecognized_keys:
-            logger.warning(f"Ignoring unrecognized keys in posedict: {unrecognized_keys}")
-        # Missing keys are fine - keys for zero values can simply be omitted.
-
-        pose = [0.0 for i in range(self.poser.get_num_parameters())]
-        for idx, key in enumerate(posedict_keys):
-            pose[idx] = posedict.get(key, 0.0)
-        return pose
 
     def on_save_image(self, event: wx.Event) -> None:
         """Ask the user for destination and save the output image.
@@ -1036,7 +889,7 @@ class MainFrame(wx.Frame):
                     current_emotion_old_index = self.emotion_choice.GetSelection()
                     current_emotion_name = self.emotion_choice.GetString(current_emotion_old_index)
 
-                    self.emotions, self.emotion_names = load_emotion_presets()
+                    self.emotions, self.emotion_names = load_emotion_presets("emotions")
                     self.emotion_choice.SetItems(self.emotion_names)
 
                     current_emotion_new_index = self.emotion_choice.FindString(current_emotion_name)
@@ -1076,13 +929,13 @@ class MainFrame(wx.Frame):
                     if emotion_name.startswith("[") and emotion_name.endswith("]"):
                         continue  # skip "[custom]" and "[reset]"
                     try:
-                        pose = self.posedict_to_pose(posedict)
+                        pose = posedict_to_pose(posedict)
 
                         posetensor = torch.tensor(pose, device=self.device, dtype=self.dtype)
                         output_index = self.output_index_choice.GetSelection()
                         with torch.no_grad():
                             output_image = self.poser.pose(self.torch_source_image, posetensor, output_index)[0].detach().cpu()
-                        numpy_image = convert_output_image_from_torch_to_numpy(output_image)
+                        numpy_image = torch_image_to_numpy(output_image)
 
                         image_file_name = os.path.join(dir_name, f"{emotion_name}.png")
                         self.save_numpy_image(numpy_image, image_file_name)
@@ -1107,11 +960,11 @@ class MainFrame(wx.Frame):
         os.makedirs(os.path.dirname(image_file_name), exist_ok=True)
         pil_image.save(image_file_name)
 
-        data_dict = self.current_pose_to_posedict()
+        pose_dict = pose_to_posedict(self.get_current_pose())
         json_file_path = os.path.splitext(image_file_name)[0] + ".json"
 
         filename_without_extension = os.path.splitext(os.path.basename(image_file_name))[0]
-        data_dict_with_filename = {filename_without_extension: data_dict}  # Create a new dict with the filename as the key
+        data_dict_with_filename = {filename_without_extension: pose_dict}  # JSON structure: {emotion_name0: posedict0, ...}
 
         try:
             with open(json_file_path, "w") as file:
